Actualizando ConsultaInteligente
