--- Codigo/Componentes/Consultas/BusquedaInteligente/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FontAwesome.Sharp" version="6.1.1" targetFramework="net472" />
-</packages>
