+++ Codigo/Componentes/Consultas/Capa_Vista/ConsultaInteligent.cs
@@ -7,6 +7,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using Capa_Controlador;
+
 using System.Data.Odbc;
 
 namespace BusquedaInteligente
@@ -17,7 +19,7 @@ namespace BusquedaInteligente
         {
             InitializeComponent();
         }
-
+       
 
         /*private void button1_Click(object sender, EventArgs e)
          {
@@ -28,12 +30,12 @@ namespace BusquedaInteligente
              }
          }
         */
-        CRUD crud = new CRUD();
+        //CRUD crud = new CRUD();
 
         private void btnNuevo_Click(object sender, EventArgs e)
         {
-            bool resultado = crud.InsertBusqueda(textBox12.Text, textBox14.Text, textBox15.Text, textBox11.Text);
-            if (resultado)
+            //bool resultado = crud.InsertBusqueda(textBox12.Text, textBox14.Text, textBox15.Text, textBox11.Text);
+            //if (resultado)
             {
                 MessageBox.Show("Datos guardados");
             }
@@ -42,19 +44,19 @@ namespace BusquedaInteligente
 
         private void iconButton7_Click(object sender, EventArgs e)
         {
-            textBox12.Clear();
-            textBox14.Clear();
-            textBox15.Clear();
-            textBox11.Clear();
+            //textBox12.Clear();
+            //textBox14.Clear();
+            //textBox15.Clear();
+            //textBox11.Clear();
 
 
         }
 
-        CRUDC crudc = new CRUDC();
+        //CRUDC crudc = new CRUDC();
         private void iconButton3_Click(object sender, EventArgs e)
         {
-            bool resultado = crudc.InsertBusquedaCompleja(textBox16.Text, textBox17.Text, textBox10.Text);
-            if (resultado)
+            //bool resultado = crudc.InsertBusquedaCompleja(textBox16.Text, textBox17.Text, textBox10.Text);
+            //if (resultado)
             {
                 MessageBox.Show("Datos guardados");
             }
@@ -63,7 +65,7 @@ namespace BusquedaInteligente
         private void iconButton8_Click(object sender, EventArgs e)
         {
             textBox16.Clear();
-            textBox17.Clear();
+            //textBox17.Clear();
             textBox10.Clear();
 
         }
@@ -78,8 +80,8 @@ namespace BusquedaInteligente
 
         private void iconButton9_Click(object sender, EventArgs e)
         {
-            textBox18.Clear();
-            textBox19.Clear();
+            //textBox18.Clear();
+            //textBox19.Clear();
             textBox9.Clear();
         }
 
@@ -95,26 +97,26 @@ namespace BusquedaInteligente
 
         private void iconButton10_Click(object sender, EventArgs e)
         {
-            textBox20.Clear();
-            textBox21.Clear();
+            //textBox20.Clear();
+            //textBox21.Clear();
             textBox8.Clear();
         }
 
         private void btnEliminar_Click(object sender, EventArgs e)
         {
-            textBox20.Clear();
-            textBox21.Clear();
-            textBox8.Clear();
-            textBox18.Clear();
-            textBox19.Clear();
-            textBox9.Clear();
-            textBox16.Clear();
-            textBox17.Clear();
-            textBox10.Clear();
-            textBox12.Clear();
-            textBox14.Clear();
-            textBox15.Clear();
-            textBox11.Clear();
+            //textBox20.Clear();
+            //textBox21.Clear();
+            //textBox8.Clear();
+            //textBox18.Clear();
+            //textBox19.Clear();
+            //textBox9.Clear();
+            //textBox16.Clear();
+            //textBox17.Clear();
+            //textBox10.Clear();
+            //textBox12.Clear();
+            //textBox14.Clear();
+            //textBox15.Clear();
+            //textBox11.Clear();
 
         }
 
@@ -135,8 +137,8 @@ namespace BusquedaInteligente
                     if (dataGridView1.Columns[e.ColumnIndex].Name == "ColumnDgvElimnar")
                     {
                         string ope = dataGridView1.Rows[e.RowIndex].Cells[0].Value.ToString();
-                        bool resultado = CRUD.EliminarBusqueda(ope);
-                        if (resultado)
+                        //bool resultado = CRUD.EliminarBusqueda(ope);
+                        //if (resultado)
                         {
                             dataGridView1.Rows.RemoveAt(e.RowIndex);
                         }
@@ -149,9 +151,41 @@ namespace BusquedaInteligente
             }
         }
 
-        private void Form1_Load(object sender, EventArgs e)
+        private void panel18_Paint(object sender, PaintEventArgs e)
         {
 
+        }
+
+        private void iconButton2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        //se necesita actualizaciones para terminar boton eliminar
+        //Diana Victores 9959-19-1471
+        public void actualizaconsultas()
+        {
+          
+        }
+
+        //boton eliminar de la forma Buscar/Eliminar - Diana Victores 9959-19-1471
+        private void iconButton12_Click(object sender, EventArgs e)
+        {
+            //cn.ejecutarconsulta(textConsultaBusqueda.Text);
+            //MessageBox.Show("Las consultas con nombre " + textConsultaBusqueda.Text + " Han sido eliminadas");
+            //actualizaconsultas();
+
+            //textConsultaBusqueda.Text = "";
+        }
+
+        private void iconButton26_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void iconButton28_Click(object sender, EventArgs e)
+        {
+            actualizaconsultas();
         }
     }
     }
