+++ Codigo/Componentes/Consultas/Capa_Vista/ConsultaInteligent.resx
